+++ src/mainmenu.h
@@ -1,7 +1,6 @@
 /*
 =========================================================================
 ---[SLSK - Main Menu functionalities header file]---
-Copyright (C) 2017 Supremist (aka supremesonicbrazil)
 This file is part of Steam Linux Swiss Knife (or SLSK for short).
 Steam Linux Swiss Knife is available under the GNU GPL v3.0 license. See the
 accompanying COPYING file for more details.
@@ -25,6 +24,7 @@ class MainMenu {
         static void LockUnlock();
         static void SwitchMode(char mode, char op);
         static void BackupScan();
+        static void PrepareScan();
         static void RestoreScan();
         static void OrganizeGameList();
         static void CheckUncheck(bool check);